refactor(state): move hash and sort functions to util
